Give spam its own command file
